+++ packages/core/src/extensions/base-props.ts
@@ -1,0 +1,10 @@
+import { Extension } from '../extension';
+import { contentPlugin } from '../plugins/content-plugin';
+
+export const BaseProps = Extension.create({
+  name: 'base-props',
+
+  addPlugins() {
+    return [contentPlugin];
+  },
+});