+++ packages/core/src/extensions/index.ts
@@ -1,3 +1,4 @@
 export * from './commands';
 export * from './drag-and-drop';
+export * from './base-props';
 export * from './base-style';