+++ packages/uikit/src/index.ts
@@ -3,3 +3,4 @@ export * from './components/button';
 export * from './components/button-group';
 export * from './components/icon-button';
 export * from './components/input';
+export * from './components/form-control';