+++ packages/extension-add-block-menu/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chamaeleon/extension-add-block-menu
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @chamaeleon/core@0.9.0
+
 ## 0.0.15
 
 ### Patch Changes