+++ packages/extension-page/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chamaeleon/extension-page
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @chamaeleon/core@0.9.0
+
 ## 0.1.1
 
 ### Patch Changes